+++ src/main/java/ga/todayOutside/src/user/models/GetUsersRes.java
@@ -5,6 +5,6 @@ import lombok.*;
 @Getter
 @AllArgsConstructor
 public class GetUsersRes {
-    private final int userId;
+    private final Long userId;
     private final String email;
 }
