+++ src/main/java/ga/todayOutside/src/user/models/PostUserRes.java
@@ -6,6 +6,6 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class PostUserRes {
-    private final int userId;
+    private final Long userId;
     private final String jwt;
 }
